测试任务两级队列调度，测试架构优化

+++ src/main/java/com/peter/controller/TestController.java
@@ -50,7 +50,7 @@ public class TestController {
 
     @GetMapping("taskSize")
     @ResponseBody
-    public Integer getTaskQueueSize(){
+    public Long getTaskQueueSize(){
         return taskUploadService.getProjectsInTask();
     }
 }
